Update packages promise example and description
I didn't cover how package_module refers to a file in `/var/cfengine/modules/packages` and
must be defined as a body package_module with the same name, because that's way beyond
the primer - most people will only use `yum` and `apt_get` anyway.  And I didn't cover how
the package_module attribute of a packages promise can be omitted if a global default is
specified in body common control, which is there out of the box on 3.10 but is commented out
in 3.7 masterfiles for backward compatibility when upgrading from 3.6.

But other than those esoteric points, I think this covers the subject matter quite thoroughly
and adequately.  :)

(Note: I am aware that my statements about upgrading and downgrading are more explicit
and clear than in the documentation.  I tested to be sure they are actually accurate statements
and that the behavior is as I expected.)

Closes #6.